Switch testPosition and algoPosition class into 2 files
