--- MCU_thur/Main.mpy
@@ -1,32 +1,0 @@
-# MicroPython code for Pyboard V1.1
-# Could be test at https://micropython.org/unicorn/
-
-import pyb
-import machine
-import time
-import USB_communication
-
-# Define pin 
-MuxS0Pin = machine.Pin('Y12')   #Real pin is "Y7"
-MuxS1Pin = machine.Pin('Y8')
-MuxS2Pin = machine.Pin('X11')
-MuxS3Pin = machine.Pin('X12')
-
-ThermiqueVinPin = pyb.ADC(machine.Pin('Y4'))    #Real pin is "X22"
-LambdaV1Pin = pyb.ADC(machine.Pin('X19'))
-LambdaV2Pin = pyb.ADC(machine.Pin('X20'))
-LambdaV3Pin = pyb.ADC(machine.Pin('X21'))
-LambdaV4Pin = pyb.ADC(machine.Pin('X3'))    #pin SYN on PCB
-
-# Test read value
-while True:
-    value = ThermiqueVinPin.read()
-    print(value)
-    time.sleep(0.05)
-    
-
-
-
-
-
-
