+++ PC/AlgoPosition.py
@@ -20,230 +20,106 @@ class AlgoPosition():
 
       position = (0,0)
       return position
-
-
-def interpolate_circle(matrix, position, radius, center=(0, 0), resolution=300, rbf_function='gaussian'):
-    """
-    Réalise l'interpolation et extrapolation des données sur une grille couvrant un cercle.
     
-    Paramètres:
-      - matrix : matrice de valeurs (par exemple, températures) de dimension (n, m)
-      - position : matrice de positions associée de dimension (n, m, 2) où chaque position = [x, y]
-      - radius : rayon du cercle d'interpolation
-      - center : centre du cercle (tuple (x, y)), par défaut (0, 0)
-      - resolution : nombre de points pour la grille d'interpolation (définit la finesse)
-      - rbf_function : type de fonction de base radiale utilisée ('multiquadric', 'linear', 'cubic', 'gaussian', etc.)
+
+    def interpolate_circle(matrix, position, radius, center=(0, 0), resolution=300, rbf_function='gaussian'):
+      """
+      Réalise l'interpolation et extrapolation des données sur une grille couvrant un cercle.
       
-    Retourne:
-      - X, Y, Z : grilles de coordonnées et valeurs interpolées (avec masquage des points hors du cercle)
-    """
-    # Extraction des données
-    x_list, y_list, z_list = [], [], []
-    n, m = matrix.shape
-    for i in range(n):
-        for j in range(m):
-            x_list.append(position[i, j, 0])
-            y_list.append(position[i, j, 1])
-            z_list.append(matrix[i, j])
-    x = np.array(x_list)
-    y = np.array(y_list)
-    z = np.array(z_list)
-    
-    # Création de l'interpolateur RBF
-    rbf = Rbf(x, y, z, function=rbf_function)
-    
-    # Définition de la grille (sur un carré couvrant le cercle)
-    xi = np.linspace(center[0] - radius, center[0] + radius, resolution)
-    yi = np.linspace(center[1] - radius, center[1] + radius, resolution)
-    X, Y = np.meshgrid(xi, yi)
-    
-    # Évaluation de l'interpolateur sur la grille
-    Z = rbf(X, Y)
-    
-    # Masquage des points situés hors du cercle
-    mask = (X - center[0])**2 + (Y - center[1])**2 > radius**2
-    Z[mask] = np.nan
-    
-    return X, Y, Z
-
-def plot_interpolation_2d(X, Y, Z, original_points=None):
-    """
-    Affiche la surface interpolée en 2D.
-    
-    Paramètres:
-      - X, Y, Z : grilles de coordonnées et valeurs interpolées
-      - original_points : tableau numpy de forme (n_points, 3) contenant [x, y, z] des données d'origine (optionnel)
-      - title : titre du graphique
-    """
-    plt.figure()
-    plt.pcolormesh(X, Y, Z, cmap='turbo', shading='auto')
-    if original_points is not None:
-        plt.scatter(original_points[:, 0], original_points[:, 1], c=original_points[:, 2], cmap='turbo', edgecolors='w', marker="s")
-    plt.colorbar(label='Température (°C)')
-    plt.title("Interpolation/extrapolation dans un cercle")
-    plt.show()
-
-
-def plot_3d(X, Y, Z, original_points=None, title='Interpolation dans le cercle'):
-    """
-    Affiche la surface interpolée en 3D.
-    
-    Paramètres:
-      - X, Y, Z : grilles de coordonnées et valeurs interpolées
-      - original_points : tableau numpy de forme (n_points, 3) contenant [x, y, z] des données d'origine (optionnel)
-      - title : titre du graphique
-    """
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection='3d')
-    surf = ax.plot_surface(X, Y, Z, cmap='turbo', edgecolor='none')
+      Paramètres:
+        - matrix : matrice de valeurs (par exemple, températures) de dimension (n, m)
+        - position : matrice de positions associée de dimension (n, m, 2) où chaque position = [x, y]
+        - radius : rayon du cercle d'interpolation
+        - center : centre du cercle (tuple (x, y)), par défaut (0, 0)
+        - resolution : nombre de points pour la grille d'interpolation (définit la finesse)
+        - rbf_function : type de fonction de base radiale utilisée ('multiquadric', 'linear', 'cubic', 'gaussian', etc.)
         
-    ax.set_title(title)
-    plt.show()
-
-def find_max_interpolation(X, Y, Z):
-    """
-    Trouve la position (x, y) et la valeur maximale dans la grille interpolée.
-    Prend en compte que certains points peuvent être masqués (np.nan).
-    
-    Retourne:
-      - max_x, max_y, max_value
-    """
-    # On utilise np.nanargmax pour ignorer les nan
-    index_max = np.nanargmax(Z)
-    max_value = np.nanmax(Z)
-    i, j = np.unravel_index(index_max, Z.shape)
-    max_x = X[i, j]
-    max_y = Y[i, j]
-    return max_x, max_y, max_value
-
-# Fonction de plot en matrice de couleur
-def plot_matrix_color(matrix):
-    plt.matshow(matrix, cmap='turbo')
-    plt.colorbar()
-    plt.show()
-
-
-# -------------------- Exemple d'utilisation --------------------
-if __name__ == '__main__':
-    """
-    Exemple d'utilisation de l'interpolation/extrapolation dans un cercle.
-
-    En premier, montre la fonction originale gaussienne en 3D.
-    Ensuite, génère des températures sur une grille 4x4 avec ajout de bruit.
-    Réalise l'interpolation/extrapolation dans un cercle de rayon 12.5 mm.
-    Plot les résultats en 2D et 3D.
-    """
-
-
-    csv_simulation_1 = "Thermique\Simulation 03-26\Offset_1_10W_parsed.csv"
-
-    # Define the grid size and spacing
-    position_xy = np.array([
-        [[-10.5, 10.5], [-3.5, 10.5], [3.5, 10.5], [10.5, 10.5]],
-        [[-10.5, 3.5],  [-3.5, 3.5],  [3.5, 3.5],  [10.5, 3.5]],
-        [[-10.5, -3.5], [-3.5, -3.5], [3.5, -3.5], [10.5, -3.5]],
-        [[-10.5, -10.5],[-3.5, -10.5],[3.5, -10.5], [10.5, -10.5]]
-    ])
+      Retourne:
+        - X, Y, Z : grilles de coordonnées et valeurs interpolées (avec masquage des points hors du cercle)
+      """
+      # Extraction des données
+      x_list, y_list, z_list = [], [], []
+      n, m = matrix.shape
+      for i in range(n):
+          for j in range(m):
+              x_list.append(position[i, j, 0])
+              y_list.append(position[i, j, 1])
+              z_list.append(matrix[i, j])
+      x = np.array(x_list)
+      y = np.array(y_list)
+      z = np.array(z_list)
+      
+      # Création de l'interpolateur RBF
+      rbf = Rbf(x, y, z, function=rbf_function)
+      
+      # Définition de la grille (sur un carré couvrant le cercle)
+      xi = np.linspace(center[0] - radius, center[0] + radius, resolution)
+      yi = np.linspace(center[1] - radius, center[1] + radius, resolution)
+      X, Y = np.meshgrid(xi, yi)
+      
+      # Évaluation de l'interpolateur sur la grille
+      Z = rbf(X, Y)
+      
+      # Masquage des points situés hors du cercle
+      mask = (X - center[0])**2 + (Y - center[1])**2 > radius**2
+      Z[mask] = np.nan
+      
+      return X, Y, Z
     
 
 
-    # Read the CSV file into a DataFrame
-    df = pd.read_csv(csv_simulation_1)
-
-    # create de X, Y arrays with the themperature values
-    x_simulation_1 = df['COORDINATES.X']
-    y_simulation_1 = df['COORDINATES.Y']
-    temp_simulation_1 = df['NDTEMP.T']
-
-    heatsink_temperature = 39
-
-    #plot the data in 3d
-    fig = plt.figure()
-    ax = fig.add_subplot(111, projection='3d')
-    ax.scatter(x_simulation_1, y_simulation_1, temp_simulation_1, c=temp_simulation_1, cmap='turbo')
-    ax.set_xlabel('X')
-    ax.set_ylabel('Y')
-    ax.set_zlabel('Temperature (°C)')
-    fig.suptitle('Temperature distribution')
-    plt.show()
-    
-
-    # transform the temp <class 'pandas.core.series.Series'> to a numpy array
-    x_simulation_1 = x_simulation_1.to_numpy()
-    y_simulation_1 = y_simulation_1.to_numpy()
-    temp_simulation_1 = temp_simulation_1.to_numpy()
-
-    #print(temp_simulation_1)
-
-    # fin the maximum of the temperature and print the position
-    temp_max = np.max(temp_simulation_1)
-    indexe_max = np.unravel_index(np.argmax(temp_simulation_1), temp_simulation_1.shape)
-    print(max(temp_simulation_1))
-    print(indexe_max[0])
-    print("position max data =", (x_simulation_1[indexe_max[0]], y_simulation_1[indexe_max[0]])  )
-    
-    # loop throuhgt the x and y arrays and return the 16 temprature where the position match the position_xy
-    # create a 4x4 matrix with the temperature values
-    temp_simulation_2 = np.zeros((4, 4))
-    for i in range(4):
-        for j in range(4):
-            # find the row where the x and y are the CLOSEST the position_xy
-            index = np.argmin(np.abs(x_simulation_1 - position_xy[i][j][0]) + np.abs(y_simulation_1 - position_xy[i][j][1]))
-            #print("index", index, "position xy", position_xy[i][j], "x", x_simulation_1[index], "y", y_simulation_1[index])
-            # get the temperature value at the index
-            temp_simulation_2[i][j] = temp_simulation_1[index]
+    def plot_interpolation_2d(X, Y, Z, original_points=None):
+      """
+      Affiche la surface interpolée en 2D.
+      
+      Paramètres:
+        - X, Y, Z : grilles de coordonnées et valeurs interpolées
+        - original_points : tableau numpy de forme (n_points, 3) contenant [x, y, z] des données d'origine (optionnel)
+        - title : titre du graphique
+      """
+      plt.figure()
+      plt.pcolormesh(X, Y, Z, cmap='turbo', shading='auto')
+      if original_points is not None:
+          plt.scatter(original_points[:, 0], original_points[:, 1], c=original_points[:, 2], cmap='turbo', edgecolors='w', marker="s")
+      plt.colorbar(label='Température (°C)')
+      plt.title("Interpolation/extrapolation dans un cercle")
+      plt.show()
 
 
-    # print the temperature matrix
-    #print("Température à chaque position :")
-    #print(position_xy)
-    #print("Température (°C) :")
-    #print(temp_simulation_2)
+    def plot_3d(X, Y, Z, original_points=None, title='Interpolation dans le cercle'):
+      """
+      Affiche la surface interpolée en 3D.
+      
+      Paramètres:
+        - X, Y, Z : grilles de coordonnées et valeurs interpolées
+        - original_points : tableau numpy de forme (n_points, 3) contenant [x, y, z] des données d'origine (optionnel)
+        - title : titre du graphique
+      """
+      fig = plt.figure()
+      ax = fig.add_subplot(111, projection='3d')
+      surf = ax.plot_surface(X, Y, Z, cmap='turbo', edgecolor='none')
+          
+      ax.set_title(title)
+      plt.show()
 
-    
-    """ 
-    # Fonction généré gaussienne
-    laser_origin = {"x":0, "y":0}
-    
-    # Fonction gaussienne centrée en (x_0, y_0) et normalisée entre 0 et 3000
-    def gaussian(x, y, x_0=0, y_0=0):
-        return 3000 * np.exp(-0.5 * ((x - x_0)**2 + (y - y_0)**2) / 10**2)
-    
-    # Affichage de la fonction continue gaussienne originale en 3D
-    X, Y = np.meshgrid(np.linspace(-rayon, rayon, 300), np.linspace(-rayon, rayon, 300))
-    Z = gaussian(X, Y, x_0=laser_origin['x'], y_0=laser_origin['y'])
-    plot_3d(X, Y, Z, title="Fonction gaussienne originale")
+    def find_max_interpolation(X, Y, Z):
+      """
+      Trouve la position (x, y) et la valeur maximale dans la grille interpolée.
+      Prend en compte que certains points peuvent être masqués (np.nan).
+      
+      Retourne:
+        - max_x, max_y, max_value
+      """
+      # On utilise np.nanargmax pour ignorer les nan
+      index_max = np.nanargmax(Z)
+      max_value = np.nanmax(Z)
+      i, j = np.unravel_index(index_max, Z.shape)
+      max_x = X[i, j]
+      max_y = Y[i, j]
+      return max_x, max_y, max_value
 
-    #maximum de la fonction gaussienne originale
-    max_x, max_y, max_value = find_max_interpolation(X, Y, Z)
-    print("Maximum de la fonction gaussienne originale à (x, y) = ({:.2f}, {:.2f}) avec une valeur de {:.2f}".format(max_x, max_y, max_value))
-
-
-    # Génération des températures sur la grille et ajout de bruit
-    Temp = gaussian(position_xy[:, :, 0], position_xy[:, :, 1], x_0=laser_origin['x'], y_0=laser_origin['y'])
-    print(Temp) """
-
-    # Prend la coubre de température dans le dossier /Thermique/Simulation 03-26/Test Lecture CSV.py
-    Temp = temp_simulation_2 - heatsink_temperature
-    print("Temps", Temp)
-
-    rayon = 30
-    noise_level = 1
-    
-    
-    # Interpolation/extrapolation dans un cercle de rayon 12.5 mm centré en (0,0)
-    X, Y, Z = interpolate_circle(Temp, position_xy, radius=rayon, center=(0, 0), resolution=100, rbf_function='gaussian')
-    
-    
-    # Affichage de la surface 
-    plot_matrix_color(Temp+heatsink_temperature)
-    plot_interpolation_2d(X, Y, Z+heatsink_temperature, original_points=np.concatenate([position_xy, Temp[:, :, None]], axis=2).reshape(-1, 3))
-    plot_3d(X, Y, Z+heatsink_temperature, title="Interpolation/extrapolation dans un cercle")
-  
-
-    # Recherche du maximum de l'interpolation
-    max_x, max_y, max_value = find_max_interpolation(X, Y, Z)
-    print("Maximum interpolé à (x, y) = ({:.2f}, {:.2f}) avec une valeur de {:.2f}".format(max_x, max_y, max_value))
-
-
-# -------------------- Fin de l'exemple --------------------
+    # Fonction de plot en matrice de couleur
+    def plot_matrix_color(matrix):
+        plt.matshow(matrix, cmap='turbo')
+        plt.colorbar()
+        plt.show()
